refactor(staking): clean old staking
BREAKING CHANGE: StakingService is obsolete
