+++ src/modules/staking/index.ts
@@ -1,2 +1,3 @@
-export * from './staking.service';
 export * from './staking-pool.service';
+export * from './staking.types';
+
